% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book_fcts.R
\name{bookfiles_get}
\alias{bookfiles_get}
\title{Copy all book files to local folder}
\usage{
bookfiles_get(path_to_copy = "~/vdr-files")
}
\arguments{
\item{path_to_copy}{Path to copy all the book files}
}
\value{
TRUE, if sucessful
}
\description{
This function will grab files from the afedR package and copy all of it to a local folder,
separated by directories including R code, data, slides and end-chapter exercises.
}
\examples{
bookfiles_get()
}
