% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exams_webex_fcts.R
\name{build_answers_text}
\alias{build_answers_text}
\title{Creates text for random questions}
\usage{
build_answers_text(text1, text2, text3)
}
\arguments{
\item{text1}{tibble 1}

\item{text2}{tibble 2}

\item{text3}{tibble 3}
}
\value{
A list
}
\description{
Creates text for random questions
}
\examples{

tibble1 <- dplyr::tibble(text = c("ABC1", "ABC2", "ABC3"), sol =  c(TRUE, FALSE, TRUE))
tibble2 <- dplyr::tibble(text = c("ABC1", "ABC2", "ABC3"), sol = c(TRUE, FALSE, TRUE))
tibble3 <- dplyr::tibble(text = c("ABC1", "ABC2", "ABC3"), sol = c(TRUE, FALSE, TRUE))
build_answers_text(tibble1, tibble2, tibble3)
}
