% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book_fcts.R
\name{colorize}
\alias{colorize}
\title{Use color in book text}
\usage{
colorize(x, color)
}
\arguments{
\item{x}{A text to be colorized}

\item{color}{the color (e.g. "red")}
}
\value{
a colorized version of string (html or latex)
}
\description{
From: https://bookdown.org/yihui/rmarkdown-cookbook/font-color.html
}
\examples{
colorize("ABC", "red")
}
