% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fcts.R
\name{data_import}
\alias{data_import}
\title{Import data from package vdr}
\usage{
data_import(name_dataset)
}
\arguments{
\item{name_dataset}{Name of the dataset filename (see \link{data_list} for more details)}
}
\value{
A path to the data file
}
\description{
This is a helper function of book "Visualização de dados com o R" by Marcelo S. Perlin.
With this function you'll be able to read the tables used in the book using only file names.
}
\examples{
file_name <- data_list()[1]
df <- data_import(file_name)
df
}
