% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fcts.R
\name{data_list}
\alias{data_list}
\title{List available datasets with description}
\usage{
data_list(be_silent = FALSE)
}
\arguments{
\item{be_silent}{Be silent?}
}
\value{
A dataframe with the data information
}
\description{
This function will list all available datasets from the book.
It also includes a description and origin, if applicable.
}
\examples{
data_list()
}
