% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercises.R
\name{exercises_compile_solution}
\alias{exercises_compile_solution}
\title{Compiles solution of exercises}
\usage{
exercises_compile_solution(dir_output = "~/afedR-solutions", run_chunks = TRUE)
}
\arguments{
\item{dir_output}{directory where to copy html file (e.g. '~/Desktop')}

\item{run_chunks}{flag to run code chunks or not. If TRUE, might take a while to process all code.}
}
\value{
path to compiled file
}
\description{
This function will compile the solution of exercises from the book to
a .html file. Aternativelly, all solutiona are available at <www.msperlin.com/afedR>
}
\examples{

\dontrun{
exercises_compile_solution(dir_output = fs::path_temp())
}
}
