% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercises.R
\name{exercises_dir_get}
\alias{exercises_dir_get}
\title{Get dir of exercise}
\usage{
exercises_dir_get(name_dir)
}
\arguments{
\item{name_dir}{name of dir}
}
\value{
full path of dir
}
\description{
Get dir of exercise
}
\examples{
exercises_dir_get(exercises_dir_list()[1])
}
