% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercises.R
\name{exercises_dir_list}
\alias{exercises_dir_list}
\title{Lists available eoc dir}
\usage{
exercises_dir_list(silent = TRUE)
}
\arguments{
\item{silent}{be silent?}
}
\value{
a char vector
}
\description{
Lists available eoc dir
}
\examples{
exercises_dir_list()
}
