% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_cash}
\alias{format_cash}
\title{Format cash}
\usage{
format_cash(x, type_cash = "USD")
}
\arguments{
\item{x}{a number}

\item{type_cash}{type of cash ("USD" or "BRL")}
}
\value{
formatted string
}
\description{
Format cash
}
\examples{
format_cash(10)
}
