% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_fct_ref}
\alias{format_fct_ref}
\title{Produces index string for pkg::function}
\usage{
format_fct_ref(pkg, this_fct, force_index = TRUE, force_pkg = FALSE)
}
\arguments{
\item{pkg}{a pkg availabe locally or github}

\item{this_fct}{name of function}

\item{force_index}{Logical (TRUE or FALSE) - defines whether to force formal
reference (e.g. [@R-pkg])}

\item{force_pkg}{force citation of package? TRUE|FALSE}
}
\value{
a string in rmarkdown
}
\description{
Formats package as `pkg` and, when called the first time, adds reference as
[@R-pkg]
}
\examples{
format_fct_ref("dplyr", "group_by")
}
