% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_fct_text}
\alias{format_fct_text}
\title{Format a function string in text}
\usage{
format_fct_text(fct)
}
\arguments{
\item{fct}{the fct name to format}
}
\value{
a char
}
\description{
Format a function string in text
}
\examples{
format_fct_text('mean')
}
