% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_number}
\alias{format_number}
\title{Formats a number (used in plots)}
\usage{
format_number(x, digits = 4)
}
\arguments{
\item{x}{a number}

\item{digits}{number of digits to use}
}
\value{
formatted number (character)
}
\description{
Formats a number (used in plots)
}
\examples{
format_number(1/3)
}
