% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_percent}
\alias{format_percent}
\title{Formats percentage}
\usage{
format_percent(x)
}
\arguments{
\item{x}{a number}
}
\value{
formatted string
}
\description{
Formats percentage
}
\examples{
format_percent(0.55)
}
