% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_pkg_citation}
\alias{format_pkg_citation}
\title{Produces package citation string}
\usage{
format_pkg_citation(
  pkg,
  github_address = NULL,
  force_ref = FALSE,
  make_index = TRUE
)
}
\arguments{
\item{pkg}{a pkg availabe locally or github}

\item{github_address}{github address of package code (eg. "msperlin/vdr")}

\item{force_ref}{Logical (TRUE or FALSE) - defines whether to force formal
reference (e.g. [@R-pkg])}

\item{make_index}{Logical (TRUE or FALSE) - defines wheter to write index for
remissive index of book}
}
\value{
a string in rmarkdown
}
\description{
Formats package as `pkg` and, when called the first time, adds reference as
[@R-pkg]
}
\examples{
format_pkg_citation("dplyr")
}
