% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_pkg_text}
\alias{format_pkg_text}
\title{Format a package string in text}
\usage{
format_pkg_text(pkg)
}
\arguments{
\item{pkg}{the pkg to format}
}
\value{
a char
}
\description{
Format a package string in text
}
\examples{
format_pkg_text("dplyr")
}
