% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_vec_as_text}
\alias{format_vec_as_text}
\title{Formats a vector of string into readable text}
\usage{
format_vec_as_text(str_in)
}
\arguments{
\item{str_in}{vector of strings}
}
\value{
a single character object
}
\description{
Formats a vector of string into readable text
}
\examples{
format_vec_as_text(c("A", "B", "C"))
}
