% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pkg_dir}
\alias{get_pkg_dir}
\title{Returns local dir of pkg}
\usage{
get_pkg_dir(sub_dir)
}
\arguments{
\item{sub_dir}{the sub dir to return path}
}
\value{
a path
}
\description{
Returns local dir of pkg
}
\examples{
get_pkg_dir("data")
}
